Share FigureManager class between gtk3 and gtk4.
The gtk version checks are not the nicest, but sharing most of the setup
code between gtk3 and gtk4 still seems worthwhile?

(Note that on gtk3, it is explicitly OK to not bother to destroy the
vbox and toolbar, as gtk3 always destroys child widgets upon parent
destruction.)
